# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 2ndShift V1 API
  version: 1.0.0
  description: High-skill IT microtask marketplace API

servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.2ndshift.com/api/v1
    description: Production server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Jobs
    description: Job management
  - name: Credits
    description: Shift Credits management
  - name: Matching
    description: Worker matching
  - name: Admin
    description: Admin operations

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, email, password, name]
              properties:
                role:
                  type: string
                  enum: [worker, client, admin]
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: User already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=...; HttpOnly; Path=/; Max-Age=604800
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string

  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'

  /jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [worker, client]
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Jobs]
      summary: Create a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    $ref: '#/components/schemas/Job'

  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    $ref: '#/components/schemas/Job'

  /jobs/{id}/apply:
    post:
      tags: [Jobs]
      summary: Apply to a job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cover_text:
                  type: string
                proposed_price:
                  type: number
                proposed_delivery:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Application created

  /credits/balance:
    get:
      tags: [Credits]
      summary: Get Shift Credits balance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance information
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
                  reserved:
                    type: integer
                  available:
                    type: integer

  /matching/auto-match:
    post:
      tags: [Matching]
      summary: Auto-match workers to a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id]
              properties:
                job_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Matched workers
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkerMatch'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [worker, client, admin]

    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, open, assigned, in_progress, completed, disputed, cancelled]
        price_fixed:
          type: number
        price_currency:
          type: string
        created_at:
          type: string
          format: date-time

    CreateJobRequest:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        microtask_id:
          type: string
          format: uuid
        price_fixed:
          type: number
        delivery_window:
          type: string
          enum: [6-24h, 3-7d, 1-4w, 1-6m]

    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer

    WorkerMatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        match_score:
          type: number
        skill_overlap:
          type: integer